+++ lib/packages/core/src/api/factories/index.ts
@@ -1,6 +1,7 @@
 /** @module core.api.factories */
 
 export * from './account';
+export * from './alias';
 export * from './block';
 export * from './network';
 export * from './message';