+++ lib/packages/core/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## 0.1.2
 - added `confirmed` property to `Account` model
 - added `alias` api
 - added `setRewardRecipient` for assigning reward recipients for miners